requirements.txt
pdfkit==0.6.1
pandas==1.1.0
boto3==1.14.31
psycopg2-binary==2.8.5